added some file functionality; haven't started with tool bar
i had some issues and screwed around with a qt generated moc file, and
its working right now but im really scared okay
one thing that is not working is ctrl+. all the other short cut things
seem to be functioning but not that one? i blame me messing with the moc
file.