balance functions work a bit better and we have cropping!
we do not account for irresponsible cropping, but we do stop cropping
after one crop. Resize does not work and I cannot fathom why. (It
crashes when it's trying to declare a function.) It upsets me greatly. I
have commented out the lines that make it break, if you could look at
them, that would be great. Also fixed up the tool bar (but we still
don't have a .qrc file, which we need). And. I can't remember what else.

Oh. Also. The extra cropFunction class is not in use, and I will dispose
of it later. But things were working, so.