Got a lot of stuff working
Icons work now.
Also photos will open and be removed.
can also move photos in the album and move around in the album.
saving and or opening files is messing up somehow but I will fix that
tomorrow,
also need to work on the edit description because thats not worked on at
all.
pretty much i am done for the night, but lots is accomplished.