got all the functionality of the buttons working correctly.
The only thing that is left to do is to set up the functionality of the
edit description, but everything else on the edit album is working
correctly including enabling and disabling of buttons. Also all save and
add functionality is working correctly. Saving photos still needs to be
accomplished.